clean up/optimize VBO code, cache color cube in PackedColor

+++ src/main/scala/li/cil/oc/Settings.scala
@@ -34,8 +34,8 @@ class Settings(val config: Config) {
   val textLinearFiltering = config.getBoolean("client.textLinearFiltering")
   val textAntiAlias = config.getBoolean("client.textAntiAlias")
   val robotLabels = config.getBoolean("client.robotLabels")
-  val soundVolume = config.getDouble("client.soundVolume").toFloat max 0 min 2
-  val fontCharScale = config.getDouble("client.fontCharScale") max 0.5 min 2
+  val soundVolume = config.getDouble("client.soundVolume").toFloat max 0.0f min 2.0f
+  val fontCharScale = config.getDouble("client.fontCharScale").toFloat max 0.5f min 2.0f
   val hologramFadeStartDistance = config.getDouble("client.hologramFadeStartDistance") max 0
   val hologramRenderDistance = config.getDouble("client.hologramRenderDistance") max 0
   val hologramFlickerFrequency = config.getDouble("client.hologramFlickerFrequency") max 0
